+++ wled00/e131.cpp
@@ -46,6 +46,7 @@ void handleE131Packet(e131_packet_t* p, IPAddress clientIP, bool isArtnet){
 
   // update status info
   realtimeIP = clientIP;
+  byte wChannel = 0;
   
   switch (DMXMode) {
     case DMX_MODE_DISABLED:
@@ -56,21 +57,25 @@ void handleE131Packet(e131_packet_t* p, IPAddress clientIP, bool isArtnet){
       if (uni != e131Universe) return;
       if (dmxChannels-DMXAddress+1 < 3) return;
       realtimeLock(realtimeTimeoutMs, mde);
+      if (realtimeOverride) return;
+      wChannel = (dmxChannels-DMXAddress+1 > 3) ? e131_data[DMXAddress+3] : 0;
       for (uint16_t i = 0; i < ledCount; i++)
-        setRealtimePixel(i, e131_data[DMXAddress+0], e131_data[DMXAddress+1], e131_data[DMXAddress+2], 0);
+        setRealtimePixel(i, e131_data[DMXAddress+0], e131_data[DMXAddress+1], e131_data[DMXAddress+2], wChannel);
       break;
 
     case DMX_MODE_SINGLE_DRGB:
       if (uni != e131Universe) return;
       if (dmxChannels-DMXAddress+1 < 4) return;
       realtimeLock(realtimeTimeoutMs, mde);
+      if (realtimeOverride) return;
+      wChannel = (dmxChannels-DMXAddress+1 > 4) ? e131_data[DMXAddress+4] : 0;
       if (DMXOldDimmer != e131_data[DMXAddress+0]) {
         DMXOldDimmer = e131_data[DMXAddress+0];
         bri = e131_data[DMXAddress+0];
         strip.setBrightness(bri);
       }
       for (uint16_t i = 0; i < ledCount; i++)
-        setRealtimePixel(i, e131_data[DMXAddress+1], e131_data[DMXAddress+2], e131_data[DMXAddress+3], 0);
+        setRealtimePixel(i, e131_data[DMXAddress+1], e131_data[DMXAddress+2], e131_data[DMXAddress+3], wChannel);
       break;
 
     case DMX_MODE_EFFECT:
@@ -103,6 +108,7 @@ void handleE131Packet(e131_packet_t* p, IPAddress clientIP, bool isArtnet){
 
     case DMX_MODE_MULTIPLE_RGB:
       realtimeLock(realtimeTimeoutMs, mde);
+      if (realtimeOverride) return;
       if (previousUniverses == 0) {
         // first universe of this fixture
         possibleLEDsInCurrentUniverse = (dmxChannels - DMXAddress + 1) / 3;
@@ -125,6 +131,7 @@ void handleE131Packet(e131_packet_t* p, IPAddress clientIP, bool isArtnet){
 
     case DMX_MODE_MULTIPLE_DRGB:
       realtimeLock(realtimeTimeoutMs, mde);
+      if (realtimeOverride) return;
       if (previousUniverses == 0) {
         // first universe of this fixture
         if (DMXOldDimmer != e131_data[DMXAddress+0]) {