+++ CHANGELOG.md
@@ -1,6 +1,50 @@
 ## WLED changelog
 
-### Development versions after 0.9.1 release
+### WLED version 0.10.0
+
+#### Build 2005030
+
+-   DMX Single RGW and Single DRGB modes now support an additional white channel
+-   Improved palettes derived from set colors and changed their names
+
+### Development versions between 0.9.1 and 0.10.0 release
+
+#### Build 2005020
+
+-   Added ACST and ACST/ACDT timezones
+
+#### Build 2005010
+
+-   Added module info page to web UI
+-   Added realtime override functionality to web UI
+-   Added individial segment power and brightness to web UI
+-   Added feature to one-click select single segment only by tapping segment name
+-   Removed palette jumping to default if color is changed
+
+#### Build 2004300
+
+-   Added realtime override option and `lor` JSON property
+-   Added `lm` (live mode) and `lip` (live IP) properties to info in JSON API
+-   Added reset commands to APIs
+-   Added `json/si`, returning state and info, but no FX or Palette lists
+-   Added rollover detection to millis(). Can track uptimes longer than 49 days
+-   Attempted to fix Wifi issues with Unifi brand APs
+
+#### Build 2004230
+
+-   Added brightness and power for individual segments
+-   Added `on` and `bri` properties to Segment object in JSON API
+-   Added `C3` an `SB` commands to HTTP get API
+-   Merged pull request #865 for 5CH_Shojo_PCB environment
+
+#### Build 2004220
+
+-   Added Candle Multi effect
+-   Added Palette capability to Pacifica effect
+
+#### Build 2004190
+
+-   Added TM1814 type LED defines
 
 #### Build 2004120